Move non-TF code in `RecoBTag/TensorFlow` to `RecoBTag/FeatureTools`
