+++ RecoBTag/FeatureTools/interface/ChargedCandidateConverter.h
@@ -1,8 +1,8 @@
-#ifndef RecoBTag_TensorFlow_ChargedCandidateConverter_h
-#define RecoBTag_TensorFlow_ChargedCandidateConverter_h
+#ifndef RecoBTag_FeatureTools_ChargedCandidateConverter_h
+#define RecoBTag_FeatureTools_ChargedCandidateConverter_h
 
-#include "RecoBTag/TensorFlow/interface/deep_helpers.h"
-#include "RecoBTag/TensorFlow/interface/TrackInfoBuilder.h"
+#include "RecoBTag/FeatureTools/interface/deep_helpers.h"
+#include "RecoBTag/FeatureTools/interface/TrackInfoBuilder.h"
 #include "DataFormats/BTauReco/interface/ChargedCandidateFeatures.h"
 
 #include "DataFormats/PatCandidates/interface/PackedCandidate.h"
@@ -10,7 +10,7 @@
 #include "DataFormats/PatCandidates/interface/Jet.h"
 
 namespace btagbtvdeep {
-  
+
   template <typename CandidateType>
     void commonCandidateToFeatures(const CandidateType * c_pf,
 				   const reco::Jet & jet,
@@ -18,7 +18,7 @@ namespace btagbtvdeep {
 				   const float & drminpfcandsv, const float & jetR,
 				   ChargedCandidateFeatures & c_pf_features,
 				   const bool flip = false) {
-    
+
 
     float trackSip2dVal = track_info.getTrackSip2dVal();
     float trackSip2dSig = track_info.getTrackSip2dSig();
@@ -33,7 +33,7 @@ namespace btagbtvdeep {
 
     c_pf_features.ptrel = catch_infs_and_bound(c_pf->pt()/jet.pt(),
 					       0,-1,0,-1);
-    
+
     c_pf_features.btagPf_trackEtaRel     =catch_infs_and_bound(track_info.getTrackEtaRel(),  0,-5,15);
     c_pf_features.btagPf_trackPtRel      =catch_infs_and_bound(track_info.getTrackPtRel(),   0,-1,4);
     c_pf_features.btagPf_trackPPar       =catch_infs_and_bound(track_info.getTrackPPar(),    0,-1e5,1e5 );
@@ -45,29 +45,29 @@ namespace btagbtvdeep {
     c_pf_features.btagPf_trackSip2dVal   =catch_infs_and_bound(trackSip2dVal, 0, -1,70 );
     c_pf_features.btagPf_trackSip2dSig   =catch_infs_and_bound(trackSip2dSig, 0, -1,4e4 );
     c_pf_features.btagPf_trackJetDistVal =catch_infs_and_bound(track_info.getTrackJetDistVal(),0,-20,1 );
-    
+
     c_pf_features.drminsv = catch_infs_and_bound(drminpfcandsv,0,-1.*jetR,0,-1.*jetR);
-    
+
   }
-  
+
   void packedCandidateToFeatures(const pat::PackedCandidate * c_pf,
 				 const pat::Jet & jet,
 				 const TrackInfoBuilder & track_info,
-				 const float drminpfcandsv, const float jetR, 
+				 const float drminpfcandsv, const float jetR,
 				 ChargedCandidateFeatures & c_pf_features,
 				 const bool flip = false) ;
-  
-  
+
+
   void recoCandidateToFeatures(const reco::PFCandidate * c_pf,
 			       const reco::Jet & jet,
 			       const TrackInfoBuilder & track_info,
 			       const float drminpfcandsv, const float jetR, const float puppiw,
 			       const int pv_ass_quality,
-			       const reco::VertexRef & pv, 
+			       const reco::VertexRef & pv,
 			       ChargedCandidateFeatures & c_pf_features,
 			       const bool flip = false) ;
-  
-  
+
+
 }
 
-#endif //RecoBTag_TensorFlow_ChargedCandidateConverter_h
+#endif //RecoBTag_FeatureTools_ChargedCandidateConverter_h
