+++ RecoBTag/FeatureTools/interface/JetConverter.h
@@ -1,15 +1,15 @@
-#ifndef RecoBTag_TensorFlow_JetConverter_h
-#define RecoBTag_TensorFlow_JetConverter_h
+#ifndef RecoBTag_FeatureTools_JetConverter_h
+#define RecoBTag_FeatureTools_JetConverter_h
 
 #include "DataFormats/BTauReco/interface/JetFeatures.h"
 
 #include "DataFormats/JetReco/interface/Jet.h"
 
 namespace btagbtvdeep {
-  
+
   class JetConverter {
   public:
-    
+
     static void jetToFeatures(const reco::Jet & jet,
 			      JetFeatures & jet_features) {
       jet_features.pt = jet.pt(); // uncorrected
@@ -17,9 +17,9 @@ namespace btagbtvdeep {
       jet_features.mass = jet.mass();
       jet_features.energy = jet.energy();
     }
-  }; 
-  
+  };
+
 }
 
-#endif //RecoBTag_TensorFlow_JetConverter_h
+#endif //RecoBTag_FeatureTools_JetConverter_h
