+++ RecoBTag/FeatureTools/interface/SecondaryVertexConverter.h
@@ -1,7 +1,7 @@
-#ifndef RecoBTag_TensorFlow_SecondaryVertexConverter_h
-#define RecoBTag_TensorFlow_SecondaryVertexConverter_h
+#ifndef RecoBTag_FeatureTools_SecondaryVertexConverter_h
+#define RecoBTag_FeatureTools_SecondaryVertexConverter_h
 
-#include "RecoBTag/TensorFlow/interface/deep_helpers.h"
+#include "RecoBTag/FeatureTools/interface/deep_helpers.h"
 #include "DataFormats/BTauReco/interface/SecondaryVertexFeatures.h"
 
 #include "DataFormats/JetReco/interface/Jet.h"
@@ -9,14 +9,14 @@
 #include "DataFormats/VertexReco/interface/Vertex.h"
 
 namespace btagbtvdeep {
-  
-  
+
+
   void svToFeatures( const reco::VertexCompositePtrCandidate & sv,
 		     const reco::Vertex & pv, const reco::Jet & jet,
 		     SecondaryVertexFeatures & sv_features,
 		     const bool flip = false) ;
-    
-  
+
+
 }
 
-#endif //RecoSV_DeepFlavour_SecondaryVertexConverter_h
+#endif //RecoBTag_FeatureTools_SecondaryVertexConverter_h
