+++ RecoBTag/FeatureTools/src/BoostedDoubleSVTagInfoConverter.cc
@@ -1,5 +1,5 @@
-#include "RecoBTag/TensorFlow/interface/BoostedDoubleSVTagInfoConverter.h"
-#include "RecoBTag/TensorFlow/interface/deep_helpers.h"
+#include "RecoBTag/FeatureTools/interface/BoostedDoubleSVTagInfoConverter.h"
+#include "RecoBTag/FeatureTools/interface/deep_helpers.h"
 #include "DataFormats/BTauReco/interface/BoostedDoubleSVTagInfoFeatures.h"
 
 #include "DataFormats/BTauReco/interface/BoostedDoubleSVTagInfo.h"
@@ -9,7 +9,7 @@ namespace btagbtvdeep {
 
   void doubleBTagToFeatures(const reco::TaggingVariableList & tag_info_vars,
 			    BoostedDoubleSVTagInfoFeatures & tag_info_features) {
-    
+
     tag_info_features.jetNTracks = tag_info_vars.get(reco::btau::jetNTracks, -999);
     tag_info_features.jetNSecondaryVertices = tag_info_vars.get(reco::btau::jetNSecondaryVertices, -999);
     tag_info_features.trackSip3dSig_0 = tag_info_vars.get(reco::btau::trackSip3dSig_0, -999);
@@ -38,9 +38,9 @@ namespace btagbtvdeep {
     tag_info_features.tau2_flightDistance2dSig = tag_info_vars.get(reco::btau::tau2_flightDistance2dSig, -999);
     tag_info_features.tau2_vertexDeltaR = tag_info_vars.get(reco::btau::tau2_vertexDeltaR, -999); // not used
     tag_info_features.z_ratio = tag_info_vars.get(reco::btau::z_ratio, -999);
-    
-  } 
-  
+
+  }
+
 }
