+++ RecoBTag/FeatureTools/src/TrackInfoBuilder.cc
@@ -1,4 +1,4 @@
-#include "RecoBTag/TensorFlow/interface/TrackInfoBuilder.h"
+#include "RecoBTag/FeatureTools/interface/TrackInfoBuilder.h"
 #include "DataFormats/PatCandidates/interface/PackedCandidate.h"
 #include "DataFormats/ParticleFlowCandidate/interface/PFCandidate.h"
 #include "DataFormats/BTauReco/interface/JetTagInfo.h"
