+++ RecoBTag/FeatureTools/src/deep_helpers.cc
@@ -1,4 +1,4 @@
-#include "RecoBTag/TensorFlow/interface/deep_helpers.h"
+#include "RecoBTag/FeatureTools/interface/deep_helpers.h"
 #include "DataFormats/PatCandidates/interface/PackedCandidate.h"
 
 namespace btagbtvdeep {
