+++ RecoBTag/TensorFlow/plugins/BuildFile.xml
@@ -4,6 +4,7 @@
   <use name="DataFormats/BTauReco"/>
   <use name="DataFormats/Common"/>
   <use name="RecoBTag/Combined"/>
+  <use name="RecoBTag/FeatureTools"/>
   <use name="RecoBTag/TensorFlow"/>
   <use name="DataFormats/PatCandidates"/>
   <use name="CommonTools/UtilAlgos"/>
