--- RecoBTag/TensorFlow/src/JetConverter.cc
@@ -1,4 +1,0 @@
-#include "RecoBTag/TensorFlow/interface/JetConverter.h"
-
-
-
